Fixed Date format during reset
previous date was "m/dd/YY" which is the Minutes followed by a zero-padded Day and Year.  Minutes was certainly incorrect, but since Dates can be confusing depending the locale when represented as a MM/DD format or DD/MM format, I adjusted to a non-zero-padded day followed by a the 3 character Month abbreviation (e.g. 6-Feb-17).